1. Create a full width page
Create a full width page with a header, a footer and a main section in
index.html. The main section should contain a content section and a
sidebar. Set the width of the content section to 70% and the sidebar
section to 30%. The sidebar should appear on the right hand side of the
content section.

index.html should include these assets:

* https://ajax.googleapis.com/ajax/libs/jquery/1.7.1/jquery.js
* test.js
* test.css

The header section should contain a navigation bar with three links:

* Home ( index.html )
* About ( /about )
* Blog ( /blog )